add skipUpdate to wm-update

+++ src/update/index.js
@@ -3,25 +3,27 @@ const prompt = require('prompt-sync')()
 const { getPackageVersion } = require('../utils/getPackageVersion')
 const fs = require('fs')
 
-const update = () => {
-  const willUpdate = prompt('Do you want to update block to major version?')
+const update = skip => {
+  if (!skip) {
+    const willUpdate = prompt('Do you want to update block to major version?')
 
-  if (willUpdate === 'y' || willUpdate === 'Y') {
-    const currentVersion = getPackageVersion()
+    if (willUpdate === 'y' || willUpdate === 'Y') {
+      const currentVersion = getPackageVersion()
 
-    fs.writeFileSync(
-      'package.json',
-      fs
-        .readFileSync('package.json')
-        .toString()
-        .replace(
-          `"version": "${currentVersion}"`,
-          `"version": "${+currentVersion.split('.')[0] + 1}.0.0"`
-        )
-    )
+      fs.writeFileSync(
+        'package.json',
+        fs
+          .readFileSync('package.json')
+          .toString()
+          .replace(
+            `"version": "${currentVersion}"`,
+            `"version": "${+currentVersion.split('.')[0] + 1}.0.0"`
+          )
+      )
+    }
   }
 }
 
 module.exports = {
-  update: () => update()
+  update: skip => update(skip)
 }
